Extract the refresh button into a component
