+++ src/scripts/shell/refresh-stores.html
@@ -1,0 +1,1 @@
+<span ng-click="$ctrl.refresh($event)" translate-attr="{ title: 'Header.Refresh'}"><i ng-class="{'fa-spin': $ctrl.loadingTracker.active()}" class="fa fa-refresh"></i></span>
